// src/interfaces/apadrinhamentoInterfaces.ts
export enum ApadrinhamentoFiltrosEnum {
    apadrinhar_com = "apadrinhar-com",
    escolher_quem_apadrinhar = "escolher-quem-apadrinhar",
}

export enum ApadrinhamentoOpcoesEnum {
    dinheiro = "dinheiro",
    racao = "racao",
    lar_temporario = "lar-temporario",
}

export enum ApadrinhamentoEscolherOpcaoEnum {
    sim = "sim",
    nao = "nao",
}